fix(permission): first init
first init should not use shared config as schema would be invalid
